requirements.txt
aiomcache==0.8.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiomcache==0.8.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.8.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.10
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langsmith==0.3.11
lxml==5.3.1
markdown2==2.5.3
MarkupSafe==3.0.2
motor==3.7.0
multidict==6.1.0
nltk==3.9.1
numpy==2.2.3
openai==1.65.1
orjson==3.10.15
packaging==24.2
pdfminer.six==20240706
pillow==11.1.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pymongo==4.11.1
pyphen==0.17.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==64.1
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
zstandard==0.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiomcache==0.8.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.8.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.10
fonttools==4.56.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langsmith==0.3.11
lxml==5.3.1
markdown2==2.5.3
MarkupSafe==3.0.2
motor==3.7.0
multidict==6.1.0
nltk==3.9.1
numpy==2.2.3
openai==1.65.1
orjson==3.10.15
packaging==24.2
pdfminer.six==20240706
pillow==11.1.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pymongo==4.11.1
pyphen==0.17.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
zstandard==0.23.0
